pyproject.toml
[project]
name = "gokay_ea"
dependencies = [
  "stk",
  "stko",
  "pywindowx",
  "atomlite",
]
requires-python = ">=3.11"
dynamic = ["version"]
readme = "README.rst"
maintainers = [
 { name = "Gokay Avci", email = "g.avci@imperial.ac.uk" },
]

[project.scripts]
gokay_ea = "gokay_ea.main:main"

[project.urls]
github = "https://github.com/gokay-avci/gokay_ea"

[project.optional-dependencies]
dev = [
  "black",
  "mypy",
  "ruff",
]

[tool.setuptools_scm]
fallback_version = "v0.0.0"

[tool.black]
line-length = 79

[tool.ruff]
line-length = 79
extend-select = ["I"]

[tool.mypy]
show_error_codes = true
implicit_optional = false
warn_no_return = true
strict_optional = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = [
  "rdkit.*",
  "pywindow.*",
  "stko.*",
]
ignore_missing_imports = true

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"
